feat(slash): 新增 /clear 命令 — 手动清除 system prompt 静态层缓存并触发重建
Source: issue #787
Type: feat
